Docs and move Retry to runtime
