+++ connect/api/src/main/java/org/apache/kafka/connect/handlers/ErrorHandlerResponse.java
@@ -5,10 +5,19 @@ package org.apache.kafka.connect.handlers;
  */
 public enum ErrorHandlerResponse {
 
+    /**
+     * retry the previous operation
+     */
     RETRY(1),
 
+    /**
+     * drop the record and move to the next one
+     */
     SKIP(2),
 
+    /**
+     * throw an Exception, and kill the task
+     */
     FAIL(3);
 
     private final int id;
