+++ connect/api/src/main/java/org/apache/kafka/connect/handlers/ProcessingContext.java
@@ -1,6 +1,6 @@
 package org.apache.kafka.connect.handlers;
 
-import java.util.Properties;
+import java.util.Map;
 
 /**
  * This object will contain all the runtime context for an error which occurs in the Connect framework while
@@ -8,14 +8,30 @@ import java.util.Properties;
  */
 public interface ProcessingContext {
 
+    /**
+     * @return which task reported this error
+     */
     String taskId();
 
+    /**
+     * @return at what stage in processing did the error happen
+     */
     Stage stageType();
 
-    String stageName();
+    /**
+     * @return description of the stage (for example, the class, position in transform chain (if applicable))
+     */
+    Map<String, Object> stageDescription();
 
-    Properties stageProperties();
+    /**
+     * @return properties used to configure this stage
+     */
+    Map<String, Object> stageProperties();
 
+    /**
+     * @return which attempt was this (first error will be 0)
+     */
     int attempt();
+
 
 }
