+++ connect/runtime/src/main/java/org/apache/kafka/connect/runtime/TransformationChain.java
@@ -21,7 +21,7 @@ import org.apache.kafka.connect.data.SchemaAndValue;
 import org.apache.kafka.connect.errors.ConnectException;
 import org.apache.kafka.connect.handlers.ErrorHandler;
 import org.apache.kafka.connect.handlers.ProcessingContext;
-import org.apache.kafka.connect.handlers.Retry;
+import org.apache.kafka.connect.runtime.handlers.Retry;
 import org.apache.kafka.connect.runtime.handlers.FailFastErrorHandler;
 import org.apache.kafka.connect.transforms.Transformation;
